Removed duplicated Position class and refactored other classes accordingly
