+++ mobile/src/main/java/com/wedriveu/mobile/service/scheduling/SchedulingServiceImpl.java
@@ -13,7 +13,6 @@ import com.wedriveu.mobile.service.ServiceExceptionHandler;
 import com.wedriveu.mobile.service.ServiceOperationCallback;
 import com.wedriveu.mobile.service.ServiceResult;
 import com.wedriveu.mobile.store.UserStore;
-import com.wedriveu.shared.rabbitmq.message.Position;
 import com.wedriveu.shared.rabbitmq.message.VehicleRequest;
 import com.wedriveu.shared.rabbitmq.message.VehicleResponse;
 import com.wedriveu.shared.rabbitmq.communication.DefaultRabbitMqCommunicationManager;
@@ -22,6 +21,7 @@ import com.wedriveu.shared.rabbitmq.communication.config.RabbitMqCommunicationCo
 import com.wedriveu.shared.rabbitmq.communication.strategy.RabbitMqCloseCommunicationStrategy;
 import com.wedriveu.shared.rabbitmq.communication.strategy.RabbitMqConsumerStrategy;
 import com.wedriveu.shared.util.Constants;
+import com.wedriveu.shared.util.Position;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
