+++ services/analytics/src/test/java/com/wedriveu/services/analytics/vehicleServiceFake/VehicleListGeneratorResponseHandler.java
@@ -5,7 +5,7 @@ import com.wedriveu.services.shared.entity.VehicleListObject;
 import com.wedriveu.services.shared.rabbitmq.VerticlePublisher;
 import com.wedriveu.services.shared.util.Log;
 import com.wedriveu.services.shared.vertx.VertxJsonMapper;
-import com.wedriveu.shared.rabbitmq.message.Position;
+import com.wedriveu.shared.util.Position;
 import io.vertx.core.eventbus.Message;
 import io.vertx.core.json.JsonObject;
