+++ services/booking/src/test/java/com/wedriveu/services/booking/entity/BookingStoreTest.java
@@ -3,7 +3,7 @@ package com.wedriveu.services.booking.entity;
 import com.wedriveu.services.shared.entity.Booking;
 import com.wedriveu.services.shared.entity.EntityListStoreStrategy;
 import com.wedriveu.services.shared.entity.JsonFileEntityListStoreStrategyImpl;
-import com.wedriveu.shared.rabbitmq.message.Position;
+import com.wedriveu.shared.util.Position;
 import org.junit.Before;
 import org.junit.Test;
