+++ services/vehicle/src/main/java/com/wedriveu/services/vehicle/boundary/nearest/VehicleFinderVerticle.java
@@ -8,8 +8,8 @@ import com.wedriveu.services.shared.rabbitmq.nearest.VehicleResponseCanDrive;
 import com.wedriveu.services.shared.util.PositionUtils;
 import com.wedriveu.services.vehicle.rabbitmq.Messages;
 import com.wedriveu.services.vehicle.rabbitmq.UserRequest;
-import com.wedriveu.shared.rabbitmq.message.Position;
 import com.wedriveu.shared.util.Constants;
+import com.wedriveu.shared.util.Position;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.eventbus.Message;
