+++ services/vehicle/src/main/java/com/wedriveu/services/vehicle/rabbitmq/UserRequest.java
@@ -2,7 +2,7 @@ package com.wedriveu.services.vehicle.rabbitmq;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.wedriveu.services.shared.entity.Vehicle;
-import com.wedriveu.shared.rabbitmq.message.Position;
+import com.wedriveu.shared.util.Position;
 
 import java.util.List;
