+++ services/vehicle/src/test/java/com/wedriveu/services/vehicle/boundary/nearest/entity/UserDataFactoryA.java
@@ -1,7 +1,7 @@
 package com.wedriveu.services.vehicle.boundary.nearest.entity;
 
 import com.wedriveu.services.vehicle.rabbitmq.UserRequest;
-import com.wedriveu.shared.rabbitmq.message.Position;
+import com.wedriveu.shared.util.Position;
 
 /**
  * Created by Marco on 02/08/2017.
