+++ services/vehicle/src/test/java/com/wedriveu/services/vehicle/boundary/vehicleregister/entity/VehicleFactoryMini.java
@@ -1,7 +1,7 @@
 package com.wedriveu.services.vehicle.boundary.vehicleregister.entity;
 
 import com.wedriveu.services.shared.entity.Vehicle;
-import com.wedriveu.shared.rabbitmq.message.Position;
+import com.wedriveu.shared.util.Position;
 
 import java.net.MalformedURLException;
 import java.net.URL;
