+++ shared/src/main/java/com/wedriveu/shared/rabbitmq/message/VehicleRequest.java
@@ -1,5 +1,7 @@
 package com.wedriveu.shared.rabbitmq.message;
 
+import com.wedriveu.shared.util.Position;
+
 /**
  * This class represents a vehicle request performed by a client to get an available vehicle.
  *
